Added embed link to GraphQL
The Anchor RSS feed doesn't provide a URL that works with their embeddable audio player. I've used a simple RegEx formula on the episode link to generate that URL and add it to GraphQL as "embedLink".

While the audio file url IS present in the RSS feed, using the embeddable audio player is probably a more common use case than implementing custom solutions.